requirements.txt
streamlit
pdf2docx
docx2pdf
pdf2image
pillow
pypandoc
pandoc